Struct smith_waterson::SmithWaterman

pub struct SmithWaterman {
    // some fields omitted
}

The SmithWaterman struct

genome_sequence: String

read_sequence: String

matrix: DMat

Methods

impl SmithWaterman

fn new(genome_sequence: String, read_sequence: String) -> SmithWaterman

Constructs a new SmithWaterman.

Examples

use smith_waterman;

let mut smitty = smith_waterman::SmithWaterman::new("ab".to_string(), "cb".to_string());

fn align(&mut self) -> (String, String)

Runs the matrix to obtain the optimum local alignment. Uses the set_matrix_loops function. returns a tuple of strings. The fist is the genome sequence the second is the read sequence.

Examples

let mut smitty = smith_waterman::SmithWaterman::new("ab".to_string(), "cb".to_string());
let alignment = smitty.align();

fn set_matrix_loops(&mut self) -> (usize, usize)

Fills the matrix with values. This uses a naïve double loop to fill the matrix.

returns a tuple of usize. The fist is the row and the second is the column of the largest value in the matrix.

This value can also be calculated out from smitty.matrix

Examples

let mut smitty = smith_waterman::SmithWaterman::new("ab".to_string(), "cb".to_string());
let max_point = smitty.set_matrix_loops();